pyproject.toml
[project]
name = "revive-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "groq>=0.37.1",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-core>=1.2.19",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "loguru>=0.7.3",
    "markupsafe<3.0.3",
    "prek>=0.3.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
]
[tool.uv.sources]
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"

[dependency-groups]
dev = [
    "black>=26.3.1",
    "ruff>=0.15.6",
]
